Block destruction 3.0
Rework the way blocks are destroyed from the ground up because of having to destroy multiple blocks when the tank is upgraded. Now instead of destroying blocks when the bullet impacts them, the bullet controlls a collider which destroys the blocks it is on top of. That way this collider is placed further from the bullet impact to destroy more blocks if necessary.